APIs/scoreboard/requirements.txt
boto3
PyJWT 